interview_agent/pyproject.toml
[project]
name = "interview-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "livekit-agents[cartesia,openai,silero]>=1.2.14",
    "openai>=2.1.0",
    "pdfplumber>=0.11.7",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
]
